requirements.txt

Django==4.0.2
djangorestframework==3.13.1
gunicorn==20.1.0
whitenoise==5.3.0
django-cors-headers==3.11.0
django-filter
